requirements.txt
openai
chromadb
unstructured
pytesseract
pdf2image
PyMuPDF
pandas
tabulate
tqdm
python-dotenv
unstructured[pdf,docx,pptx,html,email,txt]
pdfminer.six
Pillow
poppler-utils
camelot-py[cv]
numpy
requests
pi_heif-1.1.1